+++ Source/EventFlow.TestHelpers/IntegrationTest.cs
@@ -37,6 +37,7 @@ using EventFlow.Snapshots;
 using EventFlow.Snapshots.Stores;
 using EventFlow.TestHelpers.Aggregates;
 using EventFlow.TestHelpers.Aggregates.Commands;
+using EventFlow.TestHelpers.Aggregates.Sagas;
 using EventFlow.TestHelpers.Aggregates.ValueObjects;
 using EventFlow.TestHelpers.Extensions;
 using NUnit.Framework;
@@ -60,10 +61,10 @@ namespace EventFlow.TestHelpers
         [SetUp]
         public void SetUpIntegrationTest()
         {
-            var eventFlowOptions = EventFlowOptions.New
+            var eventFlowOptions = Options(EventFlowOptions.New)
                 .AddDefaults(EventFlowTestHelpers.Assembly);
 
-            Resolver = CreateRootResolver(Options(eventFlowOptions));
+            Resolver = CreateRootResolver(eventFlowOptions);
 
             AggregateStore = Resolver.Resolve<IAggregateStore>();
             EventStore = Resolver.Resolve<IEventStore>();
@@ -99,6 +100,16 @@ namespace EventFlow.TestHelpers
         {
             var pingIds = await PublishPingCommandsAsync(thingyId, 1).ConfigureAwait(false);
             return pingIds.Single();
+        }
+
+        protected Task<ThingySaga> LoadSagaAsync(ThingyId thingyId)
+        {
+            // This is specified in the ThingySagaLocator
+            var expectedThingySagaId = new ThingySagaId($"saga-{thingyId.Value}");
+
+            return AggregateStore.LoadAsync<ThingySaga, ThingySagaId>(
+                expectedThingySagaId,
+                CancellationToken.None);
         }
 
         protected async Task<IReadOnlyCollection<PingId>> PublishPingCommandsAsync(ThingyId thingyId, int count)